pyproject.toml
[tool.poetry]
name = "whiskydatabase"
version = "0.1.0"
description = ""
authors = ["Jujuware Technologies <info@jujuware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
playwright = "^1.47.0"
pyyaml = "^6.0.2"
forex-python = "^1.8"
types-beautifulsoup4 = "^4.12.0.20240907"
aiohttp = "^3.10.8"
pandas = "^2.2.3"
rapidfuzz = "^3.10.0"
scikit-learn = "^1.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
